Added transformation matrix to frame element class and a little optimization

+++ FiniteElements/Models/Contracts/IElement.cs
@@ -11,6 +11,6 @@ namespace FiniteElements.Models.Contracts
 
         List<IFrameLoad> Loads { get; }
 
-        double Number { get; set; }
+        int Number { get; set; }
     }
 }
