+++ FiniteElements/Models/Contracts/IMaterial.cs
@@ -10,6 +10,6 @@
 
         double GModule { get; }
 
-        double Number { get; set; }
+        int Number { get; set; }
     }
 }
